server/requirements.txt
blinker==1.6.2
click==8.1.6
Flask
Flask-SQLAlchemy
itsdangerous==2.1.2
Jinja2
MarkupSafe==2.1.3
psycopg2-binary
python-dotenv==1.0.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.3.6
openai==0.27.8
flasgger==0.9.7.1
